Extract module `ty.py` (#714)
